+++ src/misc/CalculatorEngineV2/UnitMulti.js
@@ -1,0 +1,8 @@
+import {Unit} from './Unit'
+
+export class UnitMulti extends Unit {
+  constructor (initObj) {
+    super(initObj)
+    this.multi = initObj.multi
+  }
+}