+++ changelog.md
@@ -3,6 +3,7 @@ vX.X.X
 
 - Added data so that NinjaCalc website appears correctly when linked to on social media, closes #136.
 - Added ability to close a calculator tab, closes #139.
+- Added support for function-based units for numerical calculator variables, closes #140.
 
 v2.0.0
 ------